requirements.txt
opentelemetry-api==1.9.1
opentelemetry-sdk==1.9.1
opentelemetry-exporter-jaeger==1.9.1
requests==2.27.1
opentelemetry-instrumentation-requests==0.28b1